+++ src/domains/matches/matches-lastest/matches-lastest-player-rating/api/react-query-api/use-get-match-player-rating.tsx
@@ -1,3 +1,5 @@
+// types에서 import
+import type { IGetMatchPlayerRatingRequest } from "../../types";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
 import { getMatchPlayerRating } from "@matches/matches-lastest/matches-lastest-player-rating/api/matches-lastest-player-rating-api";
@@ -12,13 +14,17 @@ export function useGetMatchPlayerRating({ matchId, playerId }: UseGetMatchPlayer
   const query = useSuspenseQuery({
     queryKey: [MATCHES_LASTEST_PLAYER_RATING_QUERY_KEYS.MATCH_PLAYER_RATING, matchId, playerId],
     queryFn: async () => {
-      const response = await getMatchPlayerRating({
+      const request: IGetMatchPlayerRatingRequest = {
         match_id: matchId,
         player_id: playerId,
-      });
+      };
+
+      const response = await getMatchPlayerRating(request);
+
       if (response.error) {
         throw new Error(response.error.message);
       }
+
       return response.data; // JSON 객체 반환
     },
   });