+++ src/shared/components/layout/header/buttons/logout-button.tsx
@@ -1,11 +1,11 @@
 // src/shared/components/LogoutButton.tsx
 import React from "react";
 
-import { useAuth } from "../../domains/auth/contexts/AuthContext";
+import { useAuth } from "../../../../../domains/auth/contexts/AuthContext";
 import { LogOut } from "lucide-react";
 
 const LogoutButton: React.FC = () => {
-  const { signOut, user } = useAuth();
+  const { signOut } = useAuth();
 
   const handleLogout = async () => {
     try {